Fixed crash related to bidbutton if you're not logged in
We prevented the user to bid on their own auctions, however we forgot to check if user is null. The system crash if the user is null because it does not contain 'user_id'. Fixed it by using a computed 'isOwner' containing a check if user is null, return always true (to send user to log in page) and then a check if user is owner using ternary